djangopress/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-taggit==2.1.0
sqlparse==0.4.2
